requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
antlr4-python3-runtime==4.8
argon2-cffi==20.1.0
asn1crypto==0.24.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
Automat==20.2.0
backcall==0.2.0
bleach==3.2.1
blis==0.7.4
cachetools==4.1.1
catalogue==2.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
constantly==15.1.0
contextvars==2.4
cryptography==3.3.1
cycler==0.10.0
cymem==2.0.5
Cython==0.29.21
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fastBPE==0.1.0
filelock==3.0.12
gast==0.3.3
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
Hydra==2.5
hydra-core==1.0.4
hyperlink==20.0.1
idna==2.6
idna-ssl==1.1.0
immutables==0.14
importlib-metadata==2.0.0
importlib-resources==4.1.1
incremental==17.5.0
ipykernel==5.1.1
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyter-http-over-ws==0.0.8
jupyterlab-pygments==0.1.1
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
multidict==4.7.6
murmurhash==1.0.5
nbclient==0.5.0
nbconvert==6.0.6
nbformat==4.4.0
nest-asyncio==1.4.1
nltk==3.5
notebook==6.1.4
numpy==1.18.5
oauthlib==3.1.0
omegaconf==2.0.5
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.5
pandocfilters==1.4.2
parso==0.7.1
pathy==0.3.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
preshed==3.0.5
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pydantic==1.7.3
Pygments==2.7.1
pygobject==3.26.1
PyHamcrest==2.0.2
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
pytz==2020.5
pyxdg==0.25
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
rouge-score==0.0.4
rsa==4.6
sacremoses==0.0.43
SecretStorage==2.3.1
Send2Trash==1.5.0
sentencepiece==0.1.95
six==1.15.0
smart-open==3.0.0
spacy==3.0.1
spacy-legacy==3.0.1
srsly==2.4.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
thinc==8.0.1
tokenizers==0.10.1
torch==1.7.1
torchvision==0.8.2
tornado==6.0.4
tqdm==4.55.1
traitlets==4.3.3
transformers==4.3.3
Twisted==19.2.1
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.25.10
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.0
zipp==3.2.0
zope.interface==5.2.0
